# Services/MachineLearningDemo/MachineLearningDemo.BFF/Clients/DetectionPretrainedModel/openApi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MachineLearningDemo.Detection.PretrainedModel",
    "version": "1.0"
  },
  "paths": {
    "/api/detect": {
      "put": {
        "tags": [
          "Detection"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDetectionResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Detection"
        ],
        "parameters": [
          {
            "name": "searchTerm",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageDetectionResult"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DetectedObject": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "additionalInfo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ImageDetectionResult": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "detectionTimeMilliseconds": {
            "type": "number",
            "format": "double"
          },
          "detectionSource": {
            "type": "string",
            "nullable": true
          },
          "detectedObjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetectedObject"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}